Rename screenWidthMin.js to screenDimensions.js
