+++ src/index.js
@@ -1,5 +1,5 @@
 const useStyleQueries = require('./useStyleQueries.js');
-const {screenWidthMin} = require('./queries/screenWidthMin.js');
+const {screenWidthMin} = require('./queries/screenDimensions.js');
 
 module.exports = {
   useStyleQueries,
